move some inference model classes to `models.py`
Some inference-related classes lived in `rules/base.py`, but were used across the whole inference pipeline. It makes sense to move them to a shared `models` module.
